Potential importer refined
Only sessions work for now
